Move search history rules to feature search domain
