+++ web/src/lib/features/search/domain/search-history.ts
